This is an initial commit of the Barbican puppet module
We create the manifest for the api node and add unit tests
and a basic acceptance test.  The acceptance test sets up a
basic barbican api server that runs under gunicorn and uses
the unauthenticated auth middleware.  It has been tested on
Centos 7.

More acceptance tests to be added in subsequent commits.